File Upload
CSV grabber will get the products from someone else's liquor website to make it easy to import to WooCommerce or Shopify. 
Train and interfernce are used after getting the images of each bottle you have to remove the watermark, it was cheaper making a model than using remove.bg API. I wanted to use a model to make it super accurate because using other libraries wasn't doing an ideal job.